+++ ServerChat38/src/controller/DatabaseHandlerImp.java
@@ -26,7 +26,7 @@ public class DatabaseHandlerImp implements DatabaseHandler{
     }
 
     @Override
-    public ArrayList<User> friendListHandler(int id) throws SQLException {
+    public ArrayList<User> friendListHandler(long id) throws SQLException {
         throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
     }
